consolidate isotopes
